Extract unified endpoint helpers
