+++ extensions/muxio-ext-test/src/transports/ipc.rs
@@ -1,8 +1,7 @@
+use crate::endpoint_helpers;
 use crate::test_transport::TestTransport;
-use crate::ws_helpers;
 use async_trait::async_trait;
 use interprocess::local_socket::{GenericNamespaced, ListenerOptions, ToNsName, tokio::prelude::*};
-use muxio_rpc_service::prebuffered::RpcMethodPrebuffered;
 use muxio_rpc_service_endpoint::RpcServiceEndpoint;
 use muxio_tokio_ipc_client::IpcClient;
 use muxio_tokio_ipc_server::{
@@ -34,7 +33,7 @@ impl TestTransport for IpcClient {
         let socket_name = temp_name("roundtrip");
         let server = IpcServer::new(None);
         let endpoint = server.endpoint();
-        ws_helpers::register_standard_handlers(&*endpoint).await;
+        endpoint_helpers::register_standard_handlers(&*endpoint).await;
         // Pre-register a test error handler for the roundtrip_error test
         let _ = endpoint
             .register_prebuffered(0xBAD, |_request_bytes, _ctx| async move {
@@ -91,22 +90,7 @@ impl TestTransport for IpcClient {
         let (event_tx, mut event_rx) = tokio::sync::mpsc::unbounded_channel();
         let server = IpcServer::new(Some(event_tx));
         // Register Echo on the server endpoint so client-initiated calls work
-        let server_endpoint = server.endpoint();
-        let _ = server_endpoint
-            .register_prebuffered(
-                example_muxio_rpc_service_definition::prebuffered::Echo::METHOD_ID,
-                |request_bytes, _ctx| async move {
-                    let request =
-                        example_muxio_rpc_service_definition::prebuffered::Echo::decode_request(
-                            &request_bytes,
-                        )?;
-                    example_muxio_rpc_service_definition::prebuffered::Echo::encode_response(
-                        request,
-                    )
-                    .map_err(|e| Box::new(e) as Box<dyn std::error::Error + Send + Sync>)
-                },
-            )
-            .await;
+        endpoint_helpers::register_echo_handler(&*server.endpoint()).await;
         let name = socket_name.clone();
         tokio::spawn(async move {
             let _ = server.serve(&name).await;
