+++ extensions/muxio-ext-test/src/transports/ws.rs
@@ -1,7 +1,7 @@
+use crate::endpoint_helpers;
 use crate::test_transport::TestTransport;
 use crate::ws_helpers;
 use async_trait::async_trait;
-use muxio_rpc_service::prebuffered::RpcMethodPrebuffered;
 use muxio_rpc_service_endpoint::RpcServiceEndpoint;
 use muxio_tokio_rpc_client::RpcClient;
 use muxio_tokio_rpc_server::{
@@ -26,7 +26,7 @@ impl TestTransport for RpcClient {
         let (server, host, port) = ws_helpers::setup_ws_server().await;
         // Register handlers on the SERVER endpoint (where requests are processed)
         let server_endpoint = server.endpoint();
-        ws_helpers::register_standard_handlers(&*server_endpoint).await;
+        endpoint_helpers::register_standard_handlers(&*server_endpoint).await;
         // Pre-register a test error handler for the roundtrip_error test
         let _ = server_endpoint
             .register_prebuffered(0xBAD, |_request_bytes, _ctx| async move {
@@ -93,22 +93,7 @@ impl TestTransport for RpcClient {
     ) {
         let (server, mut event_rx, host, port) = ws_helpers::setup_ws_server_with_events().await;
         // Register Echo on the server endpoint so client-initiated calls work
-        let server_endpoint = server.endpoint();
-        let _ = server_endpoint
-            .register_prebuffered(
-                example_muxio_rpc_service_definition::prebuffered::Echo::METHOD_ID,
-                |request_bytes, _ctx| async move {
-                    let request =
-                        example_muxio_rpc_service_definition::prebuffered::Echo::decode_request(
-                            &request_bytes,
-                        )?;
-                    example_muxio_rpc_service_definition::prebuffered::Echo::encode_response(
-                        request,
-                    )
-                    .map_err(|e| Box::new(e) as Box<dyn std::error::Error + Send + Sync>)
-                },
-            )
-            .await;
+        endpoint_helpers::register_echo_handler(&*server.endpoint()).await;
         let client = ws_helpers::connect_ws_client(&host, port).await;
         let endpoint = client.get_endpoint();
